% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addConceptIds.R
\name{addConceptIds}
\alias{addConceptIds}
\title{Replace the "NEW" in the label with a Concept Id}
\usage{
addConceptIds(.input, subtract = TRUE, conn)
}
\description{
Replace the "NEW" in the label with a Concept Id
}
