% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkCardinality.R
\name{checkCardinality}
\alias{checkCardinality}
\title{Check Regimen Cardinality of 1:1}
\usage{
checkCardinality(.input)
}
\description{
Does each observation have exactly 1 Regimen based on unique length?
}
