% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkIngredientCol.R
\name{checkIngredientCol}
\alias{checkIngredientCol}
\title{Add Ingredient Column}
\usage{
checkIngredientCol(.input)
}
\description{
This function adds an Ingredient column if it does not already exist and performs a QA check to make sure that all Component to Ingredient mappings are either non-existent or exactly 1.
}
