% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concept}
\alias{concept}
\title{CONCEPT}
\format{
A data frame with 123 rows and 10 variables:
\describe{
  \item{\code{concept_id}}{character Local (non-Athena) Concept Id Number}
  \item{\code{concept_name}}{character Concept Name}
  \item{\code{domain_id}}{character Domain}
  \item{\code{vocabulary_id}}{character Vocabulary}
  \item{\code{concept_class_id}}{character Concept Class}
  \item{\code{standard_concept}}{character Concept Type}
  \item{\code{concept_code}}{character Concept Code}
  \item{\code{valid_start_date}}{character Start Date}
  \item{\code{valid_end_date}}{character End Date}
  \item{\code{invalid_reason}}{character Reason concept is no longer valid}
}
}
\source{
\url{http://athena.ohdsi.org/}
}
\usage{
concept
}
\description{
HemOnc Extension Concepts in the format of a OMOP Vocabulary CONCEPT table
}
\keyword{datasets}
