% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concept_relationship}
\alias{concept_relationship}
\title{CONCEPT_RELATIONSHIP}
\format{
A data frame with 586 rows and 6 variables:
\describe{
  \item{\code{concept_id_1}}{character Concept Id 1}
  \item{\code{concept_id_2}}{character Concept Id 2}
  \item{\code{relationship_id}}{character Relationship between Concept 1 and Concept 2}
  \item{\code{valid_start_date}}{character Start Date}
  \item{\code{valid_end_date}}{character End Date}
  \item{\code{invalid_reason}}{character Reason relationship id is no longer valid}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
concept_relationship
}
\description{
HemOnc Extension Concept relationshipsin the format of a OMOP Vocabulary CONCEPT_RELATIONSHIP table
}
\keyword{datasets}
