% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{concept_synonym}
\alias{concept_synonym}
\title{CONCEPT_SYNONYM}
\format{
A data frame with 0 rows and 3 variables:
\describe{
  \item{\code{concept_id}}{character Concept Id}
  \item{\code{concept_synonym_name}}{character Concept Synonyms, also known as Alternative Labels}
  \item{\code{language_concept_id}}{character Language Concept Id}
}
}
\source{
\url{http://somewhere.important.com/}
}
\usage{
concept_synonym
}
\description{
Synonyms of HemOnc Extension Concepts in the format of a OMOP Vocabulary CONCEPT_SYNONYM table
}
\keyword{datasets}
