% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conceptIdExists.R
\name{conceptIdExists}
\alias{conceptIdExists}
\title{Does this concept_id already exist?}
\usage{
conceptIdExists(conn, new_concept_id)
}
\arguments{
\item{new_concept_id}{integer of length 1.}
}
\description{
For new concepts to be added to the extension, make sure that it already does not exist in the 'hemonc_extension.concept' table before assigning it to a new concept.
}
