% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureInput.R
\name{configureInput}
\alias{configureInput}
\title{Configure Input}
\usage{
configureInput(
  .input,
  id_col_name = "identifier",
  regimen_col_name = "CurrentRegimen",
  component_col_name = "CurrentComponent",
  ingredient_col_name = NULL
)
}
\arguments{
\item{.input}{Input dataframe of new concepts}

\item{regimen_col_name}{Name of the input column that contains the new Regimen mapping}

\item{component_col_name}{Name of the input column that contains the new Component mapping}

\item{ingredient_col_name}{Optional argument that provides the Name of the input column that contains the Ingredient mapping. Defaults to NULL if it is not provided.}
}
\value{
input dataframe with the columns renamed to standardized column names
}
\description{
This function configures the input to normalized column names of "ID", "Regimen", "Component", and if an Ingredient column is present, "Ingredient" for pointers to the correct input dataframe columns. 2 QA checkpoints stop the function: 1. That the identifier column in the input contains values that are unique to the number of rows in the input and 2. The the output contains all the identifiers that were in the input.
}
