% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constrainHemOncExtSchema.R
\name{constrainHemOncExtSchema}
\alias{constrainHemOncExtSchema}
\title{Constrain Standardized Vocabulary Tables}
\usage{
constrainHemOncExtSchema(conn)
}
\description{
The Standardized Vocabulary section in the Postgres OMOP CDM Scripts.
}
