% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertConceptTable.R
\name{convertConceptTable}
\alias{convertConceptTable}
\title{Convert the Processed New Concepts to a CONCEPT Table}
\usage{
convertConceptTable(.input)
}
\description{
Convert the Processed New Concepts to a CONCEPT Table
}
