% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertRelationshipTable.R
\name{convertRelationshipTable}
\alias{convertRelationshipTable}
\title{Convert the Processed New Concepts to a Concept Relationship Table}
\usage{
convertRelationshipTable(.input)
}
\description{
Convert the Processed New Concepts to a Concept Relationship Table
}
