% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHemOncExtSchema.R
\name{createHemOncExtSchema}
\alias{createHemOncExtSchema}
\title{Create the HemOnc Extension Schema}
\usage{
createHemOncExtSchema(conn, drop = FALSE)
}
\arguments{
\item{conn}{A Connection object.}

\item{drop}{If TRUE, will drop the `hemonc_extension` schema if it exists.}
}
\description{
Create the HemOnc Extension Schema
}
