% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddlHemOncExtSchema.R
\name{ddlHemOncExtSchema}
\alias{ddlHemOncExtSchema}
\title{DDL Standardized Vocabulary Tables}
\usage{
ddlHemOncExtSchema(conn)
}
\description{
The ddl will only execute on the condition that there are 0 tables current in the `hemonc_extension` schema.
}
