% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNewRegimenName.R
\name{extractNewRegimenName}
\alias{extractNewRegimenName}
\title{Extract the New Regimen Name from Regimen Column}
\usage{
extractNewRegimenName(.input)
}
\description{
This function extracts the content following the "NEW " prefix in the Regimen column, with the ultimate destination set to be the concept_name column in the HemOnc Extension Concept table.
}
