% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterAnyNewConcept.R
\name{filterAnyNewConcept}
\alias{filterAnyNewConcept}
\title{Filter New Concepts}
\usage{
filterAnyNewConcept(.input)
}
\description{
This function filters for any row that contains at least 1 instance of a "NEW " string
}
