% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterOutNA.R
\name{filterOutNA}
\alias{filterOutNA}
\title{Filter out Empty Rows}
\usage{
filterOutNA(.input)
}
\description{
Filter out rows where both the Regimen and Component are NA. A flag is raised if IDs are filtered out and the missing values are saved to a flag object.
}
