% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingestRelationshipTable.R
\name{ingestRelationshipTable}
\alias{ingestRelationshipTable}
\title{Ingest New Concept Relationship}
\usage{
ingestRelationshipTable(.input, conn)
}
\description{
An observation in the new concept relationship dataframe is not appended to the HemOnc Concept Relationship table if the same concept_id_1, concept_id_2, and relationship_id combination is found. Such cases are returned as `qaIngestRelationshipTable` and `qaIngestionRelationshipTable2` is the result of a followup qa where any of the rows in the input are not found in the refreshed Concept Relationship table.
}
