% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isLocalIdentifier.R
\name{isLocalIdentifier}
\alias{isLocalIdentifier}
\title{Check if an Identifier was Locally Created}
\usage{
isLocalIdentifier(identifier)
}
\description{
Returns TRUE if the identifier parses to a Datetime, FALSE if is NA, and the actual parsed output for all other data types.
}
