% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrateConceptAncestor.R,
%   R/migrateConceptDescendant.R
\name{migrateConceptAncestor}
\alias{migrateConceptAncestor}
\title{Migrate HemOnc and RxNorm Vocabularies Ancestry}
\usage{
migrateConceptAncestor(conn, source_schema)

migrateConceptAncestor(conn, source_schema)
}
\arguments{
\item{source_schema}{The schema where the main OMOP vocabularies are loaded.}
}
\description{
This function executes on the condition that there are zero rows in the concept_ancestor table in the hemonc_extension schema.

This function executes on the condition that there are zero rows in the concept_ancestor table in the hemonc_extension schema.
}
