% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separateRowsInput.R
\name{separateRowsInput}
\alias{separateRowsInput}
\title{Separate Input Rows}
\usage{
separateRowsInput(.input)
}
\description{
This function takes the input data and:
    1. Removes all carriage return symbols
    2. Separates all non-identifier rows based on a new line symbol
    3. Trims the whitespace on the resulting parsed rows, particularly important for cases when there are multiple new lines separating 2 Components, resulting in a blank Component entry
    4. Normalizes all blanks, "NA", and NA to NA_character_
    5. Filters for rows where both Regimen and Components are non-NA the non-NA Regimen and Components
}
