% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitByNewRegimenType.R
\name{splitByNewRegimenType}
\alias{splitByNewRegimenType}
\title{Split New Regimen by Type}
\usage{
splitByNewRegimenType(.input)
}
\value{
List of length 1 or 2 based on type of new Regimen.
}
\description{
All New Components would also require creating a New Regimen concept. For example a new Component Drug X would require creating a new Regimen Drug X monotherapy. Therefore, all new concepts submitted for processing can be classified at the New Regimen Level and as 1 of 2 of the following types:
    a) New Regimen only where all the components are present in HemOnc, but the combination of those Components into a Regimen is missing.
    b) New Regimen with at least 1 new Component
This function splits the input up based on whether there is a presence of a new Component in the Regimen makeup. A flag is thrown if the output is of length 1 while any other lengths other than 2 results in a brake.
}
