% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateRawData.R
\name{updateRawData}
\alias{updateRawData}
\title{Update HemOnc Extension Data in raw-data}
\usage{
updateRawData(conn)
}
\description{
This function overwrites the concept.csv, concept_relationship.csv, and concept_synonym.csv with what is currently in the database for open source distribution.
}
\keyword{internal}
