% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migrateConcept.R
\name{migrateConcept}
\alias{migrateConcept}
\title{Migrate HemOnc and RxNorm Vocabularies}
\usage{
migrateConcept(conn, source_schema)
}
\arguments{
\item{source_schema}{The schema where the main OMOP vocabularies are loaded.}
}
\description{
This function executes on the condition that there are zero rows in the concept table.
}
